requirements.txt
fastapi~=0.115.12
requests~=2.32.4
uvicorn~=0.34.3
pydantic~=2.11.5
starlette~=0.46.2
httpx~=0.28.1
python-dotenv~=1.1.0